Move auxiliary types out of main module
